updated my code so now it is working, ready to be added to the main.py
