+++ line.py
@@ -1,0 +1,86 @@
+import streamlit as st
+from datetime import date
+import datetime
+import plotly.graph_objects as go
+import yfinance as yf
+import pandas as pd
+import numpy as np
+import math
+import matplotlib.pyplot as plt
+from prophet import Prophet
+from prophet.plot import plot_plotly
+from PIL import Image
+from sklearn.linear_model import LinearRegression
+from sklearn.model_selection import train_test_split
+from sklearn import metrics 
+from sklearn.preprocessing import MinMaxScaler
+from sklearn.preprocessing import StandardScaler
+from pandas.tseries.offsets import DateOffset
+from sklearn import svm
+
+
+st.title("Linear Regression Asset Predictor")
+
+stocks = ("BTC-USD","LINK-USD","SOL-USD","MATIC-USD","MANA-USD","DOT-USD","AVAX-USD","XLM-USD","LTC-USD","XRP-USD","BNB-USD","UNI-USD","ETH-USD","ADA-USD","USDC-USD","BAT-USD")
+selected_stocks = st.selectbox("Pick a coin for prediction",stocks)
+
+@st.cache(allow_output_mutation=True)
+def load_data(ticker):
+    data = yf.download(ticker,"2010-01-01")
+    data.reset_index(inplace=True)
+    return data
+
+data = load_data(selected_stocks)
+st.subheader('Data Head')
+st.write(data.head())
+st.subheader('Data Tail')
+st.write(data.tail())
+    
+def plot_raw_data():
+    fig=go.Figure()
+    fig.add_trace(go.Scatter(x=data["Date"],y=data['Open'], name='stock_open'))
+    fig.add_trace(go.Scatter(x=data["Date"],y=data['Close'], name='stock_close'))
+    fig.layout.update(title_text='Time Series Data', xaxis_rangeslider_visible=True)
+    st.plotly_chart(fig)
+plot_raw_data()
+
+df = data
+df['returns'] = np.log(df.Close.pct_change()+1)
+def lagit (df,lags):
+    names = []
+    for i in range(1,lags +1):
+        df['Lag_'+ str(i)] = df['returns'].shift(i)
+        names.append('Lag_'+ str(i))
+    return names
+
+lagnames = lagit(df,5)
+df.dropna(inplace=True)
+
+model = LinearRegression()
+model.fit(df[lagnames],df['returns'])
+df['prediction'] = model.predict(df[lagnames])
+
+df['direction'] = [1 if i > 0 else -1 for i in df.prediction]
+df['strategy'] = df['direction'] * df['returns']
+
+x = df['returns'].cumsum()
+y = df['strategy'].cumsum()
+
+st.title('Explaining the model')
+st.write('Logistic Regression is a statistical model that is used for binary classification problems. A binary classification problem is one in which the outcome can only have two possible values, such as true or false, or 0 or 1.'
+'The basic idea behind the logistic regression model is to find the best coefficients (also known as weights or parameters) for the input features, so that the linear boundary that is generated by the model can correctly separate the data points into the two classes. The model is trained on a labeled dataset, where the outcome variable is known, and the goal is to find the coefficients that minimize the error between the predicted probabilities and the true labels.')
+
+st.title('Explaining the strategy')
+st.write('The strategy takes the percent change based on the coin of choice, then it sets windows to calculate the moving average'
+'.The next step that the strategy takes is creating a dataframe where it takes the actual returns and if these returns are greater then 0 then it gives us back 1,otherwise -1.')
+
+st.subheader('Actual Returns vs Strategy Returns')
+if st.button('Click to see Dataframe'):
+    st.write(df.head())
+
+fig = go.Figure()
+fig.add_trace(go.Scatter(x=df.index, y=x, name='Actual Returns', line=dict(color='blue')))
+fig.add_trace(go.Scatter(x=df.index, y=y, name='Strategy Returns', line=dict(color='red')))
+fig.update_layout(title='Comparison of Actual Returns and Strategy Returns',xaxis_title='Date',yaxis_title='Returns')
+if st.button('Click here to see the plot'):
+    st.plotly_chart(fig)
